Refactor slice operations to use fixed-width integer types
- Updated include statements to include "slice.h" in slice.c.
- Changed variable types from size_t to uint64_t for better consistency and clarity in slice operations.
- Modified function signatures and implementations to reflect the new variable types.
- Updated struct definitions in slice.r.h to use uint64_t for members (start, stop, step, stride) to ensure consistent size and avoid potential overflows.
- Adjusted memory allocation and deallocation in update_slicer, determine_slicer_limit, and other related functions.
- Commented out unused function resolve_slice_string and adjusted its functionality within laud_slice.
- Ensured that slice format strings are properly handled and memory is managed efficiently.

+++ c/misc/slice.r.h
@@ -20,12 +20,12 @@ extern const void *LaudSliceClass;
 #ifdef LAUD_SLICE_IMPL
 
 struct laud_dim_slice_data {
-  size_t start, stop, step, stride;
+  uint64_t start, stop, step, stride;
 };
 
 void *laud___create_slice_data_(const void *array,
                                 const char *const slice_format,
-                                size_t *new_shape, size_t *new_length);
+                                uint64_t *new_shape, uint64_t *new_length);
 
 void *laud___narray_slice_array_(const struct laud_narray *array,
                                  const struct laud_dim_slice_data *slice_data,
@@ -42,10 +42,10 @@ struct laud_slicer {
   const char *slice_format_string;
   struct laud_dim_slice_data *slice_data;
   struct laud_dim_slice_data *current_slice_data;
-  size_t *limit;
-  size_t *current_count;
-  size_t *new_shape;
-  size_t new_length;
+  uint64_t *limit;
+  uint64_t *current_count;
+  uint64_t *new_shape;
+  uint64_t new_length;
 };
 
 #endif
